+++ lib/nord.dart
@@ -1,0 +1,20 @@
+import 'dart:ui' show Color;
+
+class Nord {
+  static const Color nord0 = Color(0xFF2e3440);
+  static const Color nord1 = Color(0xFF3b4252);
+  static const Color nord2 = Color(0xFF434c5e);
+  static const Color nord3 = Color(0xFF4c566a);
+  static const Color nord4 = Color(0xFFd8dee9);
+  static const Color nord5 = Color(0xFFe5e9f0);
+  static const Color nord6 = Color(0xFFeceff4);
+  static const Color nord7 = Color(0xFF8fbcbb);
+  static const Color nord8 = Color(0xFF88c0d0);
+  static const Color nord9 = Color(0xFF81a1c1);
+  static const Color nord10 = Color(0xFF5e81ac);
+  static const Color nord11 = Color(0xFFbf616a);
+  static const Color nord12 = Color(0xFFd08770);
+  static const Color nord13 = Color(0xFFebcb8b);
+  static const Color nord14 = Color(0xFFa3be8c);
+  static const Color nord15 = Color(0xFFb48ead);
+}